Fitting Functions

MixedLRMoE.fit_exact_VI — Function

fit_exact_VI(Y, X, α_init, model, β_init, map_matrix, re_μ_list, re_Σ_list; ...)

Fit a Mixed LRMoE model with exact observations of Y using a variational algorithm.

Arguments

Y: A matrix of response.

X: A matrix of covariates.

α: A matrix of logit regression coefficients.

model: A matrix specifying the expert functions.

β_init: a matrix of regression coefficients before the random effects.

map_matrix: a matrix that maps each observation to their corresponding factor(s) in the random effect(s).

re_μ_list: a list of arrays of the variational means of random effects.

re_Σ_list: a list of arrays of the (diagonal) variational covariance matrices of random effects.

Optional Arguments

expusure: an array of numerics, indicating the time invertal over which the count data (if applicable) are collected. If nothing is provided, it is set to 1.0 for all observations. It is assumed that all continuous expert functions are not affected by exposure.

n_sims: number of simulations used to approximate the expectation of the loglikelihood/Evidence Lower Bound (ELBO).

penalty: true (default) or false, indicating whether penalty is imposed on the magnitude of parameters.

pen_α: a numeric penalty on the magnitude of logit regression coefficients. Default is 1.0.

pen_params: an array of penalty term on the magnitude of parameters of component distributions/expert functions.

α_iter_max: Maximum number of iterations when updating α. Default is 5.

ecm_iter_max: Maximum number of iterations of the ECM algorithm. Default is 200.

print_steps: true (default) or false, indicating whether intermediate updates of parameters should be logged.

Return Values

α_fit: Fitted values of logit regression coefficients α.

β_fit: Fitted values of regression coefficients β.

comp_dist: Fitted parameters of expert functions.

re_μ_list: a list of arrays of the variational means of random effects.

re_Σ_list: a list of arrays of the (diagonal) variational covariance matrices of random effects.

ll_history: a vector of the history of ELBO of the fitted model at each iteration.

iter: Number of iterations passed in the fitting function.

ll: Loglikelihood of the fitted model (with penalty on the magnitude of parameters).

ll_np: Loglikelihood of the fitted model (without penalty on the magnitude of parameters).