Gating Function with Random Effects

MixedLRMoE.LogitGating — Type

LogitGating(α, β)

Create a struct of coefficients for a gating function with α for the fixed effects, and β for the random effects.

Arguments

α: A matrix of coefficients for the fixed effects.

β: A matrix of coefficients for the random effects. If β = nothing, then the gating function does not have random effects.

MixedLRMoE.LogitGatingSim — Function

LogitGatingSim(gate, x; re_μ_list=nothing, re_Σ_list=nothing, map_matrix=nothing, check_args=true)

Simulate the gating function at x and some specifications of random effects (if provided).

Arguments

gate: A LogitGating object.

x: A matrix of covariates.

Optional arguments

re_μ_list: A list of the means of the random effects.

re_Σ_list: A list of the diagonal covariance matrices of the random effects. By model assumption, the random effects are assumed to be independent, so only a vector of the diagonal elements (i.e. the variances) is needed.

map_matrix: a matrix that maps each observation to their corresponding factor(s) in the random effect(s).

check_args: If true (default), check the validity of the arguments.

Return Values

A matrix of the log probability values of the gating function.

MixedLRMoE.LogitGatingEval — Function

LogitGatingEval(gate, x; re_list=nothing, map_matrix=nothing, check_args=true)

Evaluate the gating function at x and some specifications of random effects (if provided).

Arguments

gate: A LogitGating object.

x: A matrix of covariates.

Optional arguments

re_list: A list of the realized values of of random effects.

map_matrix: a matrix that maps each observation to their corresponding factor(s) in the random effect(s).

check_args: If true (default), check the validity of the arguments.

Return Values

A matrix of the log probability values of the gating function.

MixedLRMoE.DiagMvNormal_KL — Function

DiagMvNormal_KL(μ, Σ)

Calculate the KL divergence between a diagonal multivariate normal distribution and a standard normal distribution.

Arguments

μ: A vector of means.

Σ: A vector of the diagonal elements in the covariance matrix.